Use default reviewers for repo
The reason for this is that we can end up with reviewers that are not in the list of default reviewers list for a given project. It means that for some repositories the reviewers cannot merge and need to add an additional reviewer. 

By using the repo reviewers we mitigate this, if others are experiencing too much noise from renovate, they can remove themselves as default reviewers from some repos.