Disallow version suffixes different than -atlassian
This PR changes the pattern for some third party dependencies to ensure we stop allowing version numbers such as `0.0.0-something-atlassian-000`. The reason for this is that Renovate keeps suggesting `-vadam-cloud-atlassian` versions for velocity, which are not compatible with server platform.

This PR changes the regex to ensure that the characters before `-atlassian` are a digit, to ensure we have the format as described above.